server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
requests>=2.31.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
openpyxl>=3.1.0
pandas>=2.0.0
pywin32>=306; sys_platform == "win32"
docx2txt>=0.8
mammoth>=1.6.0
# OCR dependencies for DOC file processing
pdf2image>=1.16.0
Pillow>=10.0.0
pytesseract>=0.3.10
python-pptx>=0.6.21
google-genai>=0.3.0


fpdf>=1.7.2
